+++ src/tile.rs
@@ -27,7 +27,7 @@ pub struct TileData {
     /// The animation frames of this tile.
     pub animation: Option<Vec<Frame>>,
     /// The type of this tile.
-    pub tile_type: Option<String>,
+    pub user_type: Option<String>,
     /// The probability of this tile.
     pub probability: f32,
 }
@@ -67,15 +67,16 @@ impl TileData {
         reader: &mut impl ResourceReader,
         cache: &mut impl ResourceCache,
     ) -> Result<(TileId, TileData)> {
-        let ((tile_type, probability), id) = get_attrs!(
+        let ((user_type, user_class, probability), id) = get_attrs!(
             for v in attrs {
-                Some("type") => tile_type ?= v.parse(),
+                Some("type") => user_type ?= v.parse(),
+                Some("class") => user_class ?= v.parse(),
                 Some("probability") => probability ?= v.parse(),
                 "id" => id ?= v.parse::<u32>(),
             }
-            ((tile_type, probability), id)
+            ((user_type, user_class, probability), id)
         );
-
+        let user_type = user_type.or(user_class);
         let mut image = Option::None;
         let mut properties = HashMap::new();
         let mut objectgroup = None;
@@ -107,7 +108,7 @@ impl TileData {
                 properties,
                 collision: objectgroup,
                 animation,
-                tile_type,
+                user_type,
                 probability: probability.unwrap_or(1.0),
             },
         ))