+++ src/tileset.rs
@@ -10,7 +10,7 @@ use crate::tile::TileData;
 use crate::{util::*, Gid, ResourceCache, ResourceReader, Tile, TileId};
 
 mod wangset;
-pub use wangset::{WangColor, WangId, WangSet, WangTile};
+pub use wangset::*;
 
 /// A collection of tiles for usage in maps and template objects.
 ///
@@ -42,6 +42,10 @@ pub struct Tileset {
     /// calculated using [image](Self::image) width, [tile width](Self::tile_width),
     /// [spacing](Self::spacing) and [margin](Self::margin).
     pub columns: u32,
+    /// The x-offset to be used when drawing tiles of this tileset.
+    pub offset_x: i32,
+    /// The y-offset to be used when drawing tiles of this tileset.
+    pub offset_y: i32,
 
     /// A tileset can either:
     /// * have a single spritesheet `image` in `tileset` ("regular" tileset);
@@ -60,6 +64,9 @@ pub struct Tileset {
 
     /// The custom properties of the tileset.
     pub properties: Properties,
+
+    /// The custom tileset type, arbitrarily set by the user.
+    pub user_type: Option<String>,
 }
 
 pub(crate) enum EmbeddedParseResultType {
@@ -79,6 +86,7 @@ struct TilesetProperties {
     tilecount: u32,
     columns: Option<u32>,
     name: String,
+    user_type: Option<String>,
     tile_width: u32,
     tile_height: u32,
     /// The root all non-absolute paths contained within the tileset are relative to.
@@ -125,18 +133,24 @@ impl Tileset {
         reader: &mut impl ResourceReader,
         cache: &mut impl ResourceCache,
     ) -> Result<EmbeddedParseResult> {
-        let ((spacing, margin, columns, name), (tilecount, first_gid, tile_width, tile_height)) = get_attrs!(
+        let (
+            (spacing, margin, columns, name, user_type, user_class),
+            (tilecount, first_gid, tile_width, tile_height),
+        ) = get_attrs!(
            for v in attrs {
             Some("spacing") => spacing ?= v.parse(),
             Some("margin") => margin ?= v.parse(),
             Some("columns") => columns ?= v.parse(),
             Some("name") => name = v,
+            Some("type") => user_type ?= v.parse(),
+            Some("class") => user_class ?= v.parse(),
+
             "tilecount" => tilecount ?= v.parse::<u32>(),
             "firstgid" => first_gid ?= v.parse::<u32>().map(Gid),
             "tilewidth" => tile_width ?= v.parse::<u32>(),
             "tileheight" => tile_height ?= v.parse::<u32>(),
            }
-           ((spacing, margin, columns, name), (tilecount, first_gid, tile_width, tile_height))
+           ((spacing, margin, columns, name, user_type, user_class), (tilecount, first_gid, tile_width, tile_height))
         );
 
         let root_path = path.parent().ok_or(Error::PathIsNotFile)?.to_owned();
@@ -147,6 +161,7 @@ impl Tileset {
                 spacing,
                 margin,
                 name: name.unwrap_or_default(),
+                user_type: user_type.or(user_class),
                 root_path,
                 columns,
                 tilecount,
@@ -189,18 +204,23 @@ impl Tileset {
         reader: &mut impl ResourceReader,
         cache: &mut impl ResourceCache,
     ) -> Result<Tileset> {
-        let ((spacing, margin, columns, name), (tilecount, tile_width, tile_height)) = get_attrs!(
+        let (
+            (spacing, margin, columns, name, user_type, user_class),
+            (tilecount, tile_width, tile_height),
+        ) = get_attrs!(
             for v in attrs {
                 Some("spacing") => spacing ?= v.parse(),
                 Some("margin") => margin ?= v.parse(),
                 Some("columns") => columns ?= v.parse(),
                 Some("name") => name = v,
+                Some("type") => user_type ?= v.parse(),
+                Some("class") => user_class ?= v.parse(),
 
                 "tilecount" => tilecount ?= v.parse::<u32>(),
                 "tilewidth" => tile_width ?= v.parse::<u32>(),
                 "tileheight" => tile_height ?= v.parse::<u32>(),
             }
-            ((spacing, margin, columns, name), (tilecount, tile_width, tile_height))
+            ((spacing, margin, columns, name, user_type, user_class), (tilecount, tile_width, tile_height))
         );
 
         let root_path = path.parent().ok_or(Error::PathIsNotFile)?.to_owned();
@@ -211,6 +231,7 @@ impl Tileset {
                 spacing,
                 margin,
                 name: name.unwrap_or_default(),
+                user_type: user_type.or(user_class),
                 root_path,
                 columns,
                 tilecount,
@@ -232,10 +253,15 @@ impl Tileset {
         let mut tiles = HashMap::with_capacity(prop.tilecount as usize);
         let mut properties = HashMap::new();
         let mut wang_sets = Vec::new();
+        let mut offset = (0i32, 0i32);
 
         parse_tag!(parser, "tileset", {
             "image" => |attrs| {
                 image = Some(Image::new(parser, attrs, &prop.root_path)?);
+                Ok(())
+            },
+            "tileoffset" => |attrs| {
+                offset = parse_tileoffset(attrs)?;
                 Ok(())
             },
             "properties" => |_| {
@@ -272,11 +298,14 @@ impl Tileset {
 
         Ok(Tileset {
             name: prop.name,
+            user_type: prop.user_type,
             tile_width: prop.tile_width,
             tile_height: prop.tile_height,
             spacing,
             margin,
             columns,
+            offset_x: offset.0,
+            offset_y: offset.1,
             tilecount: prop.tilecount,
             image,
             tiles,
@@ -300,4 +329,15 @@ impl Tileset {
                 )
             })
     }
+}
+
+/// Parse the optional <tileoffset x=... y=.../> tag.
+fn parse_tileoffset(attrs: Vec<OwnedAttribute>) -> Result<(i32, i32)> {
+    Ok(get_attrs!(
+        for v in attrs {
+            "x" => offset_x ?= v.parse::<i32>(),
+            "y" => offset_y ?= v.parse::<i32>(),
+        }
+        (offset_x, offset_y)
+    ))
 }